another (debug) commit of the upload branch integrated fixes to the http class
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/new-upload@49850 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ phase3/includes/Import.php
@@ -354,15 +354,16 @@ class WikiRevision {
 
 		// @fixme!
 		$src = $this->getSrc();
-		$data = Http::get( $src );
-		if( !$data ) {
-			wfDebug( "IMPORT: couldn't fetch source $src\n" );
-			fclose( $f );
-			unlink( $tempo );
-			return false;
+		$status = Http::get( $src );
+		if( !$status->isOK() ){		
+			if( !$data ) {
+				wfDebug( "IMPORT: couldn't fetch source $src\n" );
+				fclose( $f );
+				unlink( $tempo );
+				return false;
+			}
 		}
-
-		fwrite( $f, $data );
+		fwrite( $f, $status->value );
 		fclose( $f );
 
 		return $tempo;
